made in C language
1)Code is basically been made in C language.
2)In this program, we have used linklist.
Basically  we stored all the file content in the linklist in the insert
function.
We have used linked list so that we could store any large file without
re-adjusting the size of the datatype.
3)In the string function thta we stored quatret one by one in an array
and send it to the compare function till we do not get the end of file
(EOF).
4)In the compare function, we compared the array which we send from the
string function from each every quartet which was formed in the content
of file. If we find the repeatition then
we increment the count variable.
5)In the result_compare function, we send the count variable with the
count of the previous quartet, and if both count are equal than we
change the value of count an store the new quartet in the result array.
if both count are equal and the quartet are different than we store new
quartet in the eq_result two dimentional array.
6)When all the quartet in string function are compaired than with the
help of display function, we display the most repeating quartet.